pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stackit-audit"
version = "0.1.0"
description = "Lightweight STACKIT cloud audit tool: discovery, CCM/C5 readiness checks, JSON/Markdown/PDF reports"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "STACKIT Audit Contributors" }]
dependencies = [
    "pydantic>=2.5",
    "httpx>=0.25",
    "PyJWT[crypto]>=2.8",
    "cryptography>=41",
    "PyYAML>=6.0",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-cov>=4"]

[project.scripts]
stackit-audit = "stackit_audit.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["stackit_audit*"]

[tool.setuptools.package-data]
stackit_audit = ["frameworks/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
